Created generic KeyFrames class based on g2d Animation class.
This extends the functionality of the Animation class to objects
other than TextureRegion by introducing a new class KeyFrames<T>. The
Animation class remains as an implementation of KeyFrames<TextureRegion>.
